move files and local tests pass
