+++ sklearn/ensemble/_voting.py
