+++ sklearn/ensemble/tests/test_iforest.py
@@ -19,7 +19,7 @@ from sklearn.utils.testing import assert_allclose
 
 from sklearn.model_selection import ParameterGrid
 from sklearn.ensemble import IsolationForest
-from sklearn.ensemble.iforest import _average_path_length
+from sklearn.ensemble._iforest import _average_path_length
 from sklearn.model_selection import train_test_split
 from sklearn.datasets import load_boston, load_iris
 from sklearn.utils import check_random_state
@@ -289,7 +289,7 @@ def test_iforest_warm_start():
 # mock get_chunk_n_rows to actually test more than one chunk (here one
 # chunk = 3 rows:
 @patch(
-    "sklearn.ensemble.iforest.get_chunk_n_rows",
+    "sklearn.ensemble._iforest.get_chunk_n_rows",
     side_effect=Mock(**{"return_value": 3}),
 )
 @pytest.mark.parametrize(
@@ -304,7 +304,7 @@ def test_iforest_chunks_works1(
 
 # idem with chunk_size = 5 rows
 @patch(
-    "sklearn.ensemble.iforest.get_chunk_n_rows",
+    "sklearn.ensemble._iforest.get_chunk_n_rows",
     side_effect=Mock(**{"return_value": 10}),
 )
 @pytest.mark.parametrize(
