+++ sklearn/ensemble/tests/test_weight_boosting.py
@@ -12,7 +12,7 @@ from sklearn.model_selection import train_test_split
 from sklearn.model_selection import GridSearchCV
 from sklearn.ensemble import AdaBoostClassifier
 from sklearn.ensemble import AdaBoostRegressor
-from sklearn.ensemble import weight_boosting
+from sklearn.ensemble._weight_boosting import _samme_proba
 from scipy.sparse import csc_matrix
 from scipy.sparse import csr_matrix
 from scipy.sparse import coo_matrix
@@ -64,7 +64,7 @@ def test_samme_proba():
             return probs
     mock = MockEstimator()
 
-    samme_proba = weight_boosting._samme_proba(mock, 3, np.ones_like(probs))
+    samme_proba = _samme_proba(mock, 3, np.ones_like(probs))
 
     assert_array_equal(samme_proba.shape, probs.shape)
     assert np.isfinite(samme_proba).all()
